+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Change Log
 ==========
 
+2017-06-24
+----------
+
+ * added hasKey, hasValue and values to AbstractEnumeration
+
 2017-06-23
 ----------